*) Changes from Dre - Removed: lot of useless casts (mainly related to GameNPC.eFlags - Changed: removed useless t/c and added new ones
*) Changes from me:
- Changed: the trainer window call is now in GameTrainer class
- Removed: all class call to ttainer window
- Changed: missingg experience and freelevel 'll only be given by your class trainer or an universal trainer.

Theses changes are aiming the 1.105 patch. For your universal trainer, using eCharacter.Unknown will always call the trainer window, while others classes trainers will only work with corresponding player class.

git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk/GameServer@2592 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ commands/admincommands/plvl.cs
@@ -169,7 +169,7 @@ namespace DOL.GS.Commands
 
 						foreach( GameNPC npc in client.Player.GetNPCsInRadius( WorldMgr.VISIBILITY_DISTANCE ) )
 									{
-							if( ( npc.Flags & (int)GameNPC.eFlags.CANTTARGET ) != 0 || ( npc.Flags & (int)GameNPC.eFlags.DONTSHOWNAME ) != 0 )
+							if( ( npc.Flags & GameNPC.eFlags.CANTTARGET ) != 0 || ( npc.Flags & GameNPC.eFlags.DONTSHOWNAME ) != 0 )
 										{
 								client.Out.SendNPCCreate( npc );
 										}
