+++ spells/Artifacts/bracerofzo.cs
@@ -66,7 +66,7 @@ namespace DOL.GS.Spells
                 deamons[i].Realm = Caster.Realm;
                 deamons[i].CurrentSpeed = 0;
                 deamons[i].Level = 36;
-                deamons[i].Flags |= (uint)GameNPC.eFlags.FLYING;
+                deamons[i].Flags |= GameNPC.eFlags.FLYING;
                 deamons[i].AddToWorld();
 				(deamons[i].Brain as IOldAggressiveBrain).AddToAggroList(Caster.TargetObject as GameLiving, 1);
 				(deamons[i].Brain as ProcPetBrain).Think();
